Conflicts:
	base/siteorigin-widget.class.php

+++ base/siteorigin-widget.class.php
@@ -10,8 +10,17 @@ abstract class SiteOrigin_Widget extends WP_Widget {
 	protected $base_folder;
 	protected $repeater_html;
 	protected $field_ids;
+
+	/**
+	 * @var array The array of registered frontend scripts
+	 */
 	protected $frontend_scripts = array();
+
+	/**
+	 * @var array The array of registered frontend styles
+	 */
 	protected $frontend_styles = array();
+
 	protected $generated_css = array();
 
 	protected $current_instance;
@@ -48,7 +57,16 @@ abstract class SiteOrigin_Widget extends WP_Widget {
 	}
 
 	/**
+	 * Initialize this widget in whatever way we need to. Run before rendering widget or form.
+	 */
+	function initialize(){
+
+	}
+
+	/**
 	 * Get the form options and allow child widgets to modify that form.
+	 *
+	 * @param bool $parent
 	 *
 	 * @return mixed
 	 */
@@ -58,7 +76,10 @@ abstract class SiteOrigin_Widget extends WP_Widget {
 			$form_options = $parent->modify_child_widget_form( $form_options, $this );
 		}
 
-		return apply_filters( 'siteorigin_widget_form_options', $form_options );
+		// Give other plugins a way to modify this form.
+		$form_options = apply_filters( 'siteorigin_widgets_form_options', $form_options, $this );
+		$form_options = apply_filters( 'siteorigin_widgets_form_options_' . $this->id_base, $form_options, $this );
+		return $form_options;
 	}
 
 	/**
@@ -69,6 +90,10 @@ abstract class SiteOrigin_Widget extends WP_Widget {
 	 */
 	public function widget( $args, $instance ) {
 		$instance = $this->modify_instance($instance);
+
+		// Filter the instance
+		$instance = apply_filters( 'siteorigin_widgets_instance', $instance, $this );
+		$instance = apply_filters( 'siteorigin_widgets_instance_' . $this->id_base, $instance, $this );
 
 		$args = wp_parse_args( $args, array(
 			'before_widget' => '',
@@ -82,9 +107,9 @@ abstract class SiteOrigin_Widget extends WP_Widget {
 
 		$css_name = $this->generate_and_enqueue_instance_styles( $instance );
 		$this->enqueue_frontend_scripts( $instance );
-		extract( apply_filters( 'siteorigin_widget_template_variables', $this->get_template_variables($instance, $args) ) );
+		extract( apply_filters( 'siteorigin_widgets_template_variables', $this->get_template_variables($instance, $args) ) );
 
-		// Storage hash allows
+		// Storage hash allows templates to get access to
 		$storage_hash = '';
 		if( !empty($this->widget_options['instance_storage']) ) {
 			$stored_instance = $this->filter_stored_instance($instance);
@@ -95,15 +120,32 @@ abstract class SiteOrigin_Widget extends WP_Widget {
 			}
 		}
 
+		$template_file = siteorigin_widget_get_plugin_dir_path( $this->id_base ) . $this->get_template_dir( $instance ) . '/' . $this->get_template_name( $instance ) . '.php';
+		$template_file = apply_filters('siteorigin_widgets_template_file_' . $this->id_base, $template_file, $instance, $this );
+		$template_file = realpath($template_file);
+
+		// Don't accept non PHP files
+		if( substr($template_file, -4) != '.php' ) $template_file = false;
+
+		// Exit if the template file is empty or doesn't exist
+		if( empty($template_file) || !file_exists($template_file) ) return;
+
 		echo $args['before_widget'];
 		echo '<div class="so-widget-'.$this->id_base.' so-widget-'.$css_name.'">';
 		ob_start();
-		@ include siteorigin_widget_get_plugin_dir_path( $this->id_base ) . '/' . $this->get_template_dir( $instance ) . '/' . $this->get_template_name( $instance ) . '.php';
-		echo apply_filters('siteorigin_widget_template', ob_get_clean(), get_class($this), $instance, $this );
+		@ include $template_file;
+		echo apply_filters('siteorigin_widgets_template', ob_get_clean(), get_class($this), $instance, $this );
 		echo '</div>';
 		echo $args['after_widget'];
 	}
 
+	/**
+	 * Generate the CSS for this widget and display it in the appropriate way
+	 *
+	 * @param $instance The instance array
+	 *
+	 * @return string The CSS name
+	 */
 	function generate_and_enqueue_instance_styles( $instance ) {
 
 		$this->current_instance = $instance;
@@ -164,6 +206,13 @@ abstract class SiteOrigin_Widget extends WP_Widget {
 		return array();
 	}
 
+	/**
+	 * Render a sub widget. This should be used inside template files.
+	 *
+	 * @param $class
+	 * @param $args
+	 * @param $instance
+	 */
 	public function sub_widget($class, $args, $instance){
 		if(!class_exists($class)) return;
 		$widget = new $class;
@@ -454,11 +503,10 @@ abstract class SiteOrigin_Widget extends WP_Widget {
 		$style_name = $this->get_style_name($instance);
 		if( empty($style_name) ) return '';
 
-		$file_path = siteorigin_widget_get_plugin_dir_path( $this->id_base ).'styles/'.$style_name . '.less';
+		$less_file = siteorigin_widget_get_plugin_dir_path( $this->id_base ).'styles/'.$style_name . '.less';
+		$less_file = apply_filters( 'siteorigin_widgets_less_file_' . $this->id_base, $less_file, $instance, $this );
 
-		if ( !file_exists( $file_path ) ) return '';
-
-		$less = file_get_contents( $file_path );
+		$less = ( substr( $less_file, -5 ) == '.less' && file_exists($less_file) ) ? file_get_contents( $less_file ) : '';
 
 		// Substitute the variables
 		if( !class_exists('SiteOrigin_Widgets_Color_Object') ) require plugin_dir_path( __FILE__ ) . 'inc/color.php';
@@ -478,7 +526,8 @@ abstract class SiteOrigin_Widget extends WP_Widget {
 			}
 		}
 
-		$less = apply_filters( 'siteorigin_widget_styles', $less, get_class($this), $instance );
+		$less = apply_filters( 'siteorigin_widgets_styles', $less, get_class($this), $instance );
+		$less = apply_filters( 'siteorigin_widgets_less_' . $this->id_base, $less, $instance, $this );
 
 		$style = $this->get_style_name( $instance );
 		$hash = $this->get_style_hash( $instance );
@@ -496,13 +545,32 @@ abstract class SiteOrigin_Widget extends WP_Widget {
 		$lc_functions = new SiteOrigin_Widgets_Less_Functions($this, $instance);
 		$lc_functions->registerFunctions($c);
 
-		return apply_filters( 'siteorigin_widget_instance_css', $c->compile( $less ), $instance );
+		return apply_filters( 'siteorigin_widgets_instance_css', $c->compile( $less ), $instance, $this );
 	}
 
+	/**
+	 * Replaces LESS imports with the content from the actual files. This used as a preg callback.
+	 *
+	 * @param $matches
+	 *
+	 * @return string
+	 */
 	private function get_less_import_contents($matches) {
-		$fileName = $matches[1];
-		//get file extenstion
-		preg_match( '/\.\w+$/', $fileName, $ext );
+		$filename = $matches[1];
+
+		// First, we'll deal with a few special cases
+		switch( $filename ) {
+			case 'mixins':
+				return file_get_contents( plugin_dir_path( __FILE__ ) . 'less/mixins.less' );
+				break;
+
+			case 'lesshat':
+				return file_get_contents( plugin_dir_path( __FILE__ ) . 'less/lesshat.less' );
+				break;
+		}
+
+		//get file extension
+		preg_match( '/\.\w+$/', $filename, $ext );
 		//if there is a file extension and it's not .less or .css we ignore
 		if ( ! empty( $ext ) ) {
 			if ( ! ( $ext[0] == '.less' || $ext[0] == '.css' ) ) {
@@ -510,17 +578,17 @@ abstract class SiteOrigin_Widget extends WP_Widget {
 			}
 		}
 		else {
-			$fileName .= '.less';
+			$filename .= '.less';
 		}
 		//first check local widget styles directory and then bundle less directory
-		$searchPath = array(
+		$search_path = array(
 			siteorigin_widget_get_plugin_dir_path( $this->id_base ) . 'styles/',
 			plugin_dir_path( __FILE__ ) . 'less/'
 		);
 
-		foreach ( $searchPath as $dir ) {
-			if ( file_exists( $dir . $fileName ) ) {
-				return file_get_contents( $dir . $fileName )."\n\n";
+		foreach ( $search_path as $dir ) {
+			if ( file_exists( $dir . $filename ) ) {
+				return file_get_contents( $dir . $filename )."\n\n";
 			}
 		}
 
@@ -528,6 +596,13 @@ abstract class SiteOrigin_Widget extends WP_Widget {
 		return '';
 	}
 
+	/**
+	 * Used as a preg callback to replace .widget-function('some_function', ...) with output from less_some_function($instance, $args).
+	 *
+	 * @param $matches
+	 *
+	 * @return mixed|string
+	 */
 	private function less_widget_inject($matches){
 		// We're going to lazily split the arguments by comma
 		$args = explode(',', $matches[1]);
@@ -537,6 +612,7 @@ abstract class SiteOrigin_Widget extends WP_Widget {
 		$func = 'less_' . trim( array_shift($args) , '\'"');
 		if( !method_exists($this, $func) ) return '';
 
+		// Finally call the function and include the
 		$args = array_map('trim', $args);
 		return call_user_func( array($this, $func), $this->current_instance, $args );
 	}
@@ -649,19 +725,25 @@ abstract class SiteOrigin_Widget extends WP_Widget {
 			}
 		}
 
+		// Also let other plugins also sanitize the instance
+		$instance = apply_filters( 'siteorigin_widgets_sanitize_instance', $instance, $fields, $this );
 		return $instance;
 	}
 
 	/**
+	 * Utility function to get a field name for a widget field.
+	 *
 	 * @param $field_name
 	 * @param array $repeater
 	 * @param string $repeater_append
 	 * @return mixed|string
 	 */
-	public function so_get_field_name($field_name, $repeater = array(), $repeater_append = '[]') {
-		if( empty($repeater) ) return $this->get_field_name($field_name);
+	protected function so_get_field_name($field_name, $repeater = array(), $repeater_append = '[]') {
+		if( empty($repeater) ) {
+			return $this->get_field_name($field_name);
+		}
 		else {
-
+			// We also need to add the repeater fields
 			$repeater_extras = '';
 			foreach($repeater as $r) {
 				$repeater_extras .= '['.$r.'][#'.$r.'#]';
@@ -683,7 +765,9 @@ abstract class SiteOrigin_Widget extends WP_Widget {
 	 * @return string
 	 */
 	public function so_get_field_id( $field_name, $repeater = array(), $is_template = false ) {
-		if( empty($repeater) ) return $this->get_field_id($field_name);
+		if( empty($repeater) ) {
+			return $this->get_field_id($field_name);
+		}
 		else {
 			$name = $repeater;
 			$name[] = $field_name;
@@ -876,12 +960,8 @@ abstract class SiteOrigin_Widget extends WP_Widget {
 					<a href="#" class="media-upload-button" data-choose="<?php echo esc_attr($choose_title) ?>" data-update="<?php echo esc_attr( $update_button ) ?>" data-library="<?php echo esc_attr($library) ?>">
 						<?php echo esc_html($choose_title) ?>
 					</a>
-
-					<a href="#" class="media-remove-button"><?php esc_html_e('Remove', 'siteorigin-widgets') ?></a>
-
 				</div>
-
-				<div class="clear"></div>
+				<a href="#" class="media-remove-button <?php if( empty($value) ) echo 'remove-hide'; ?>"><?php esc_html_e('Remove', 'siteorigin-widgets') ?></a>
 
 				<?php
 				if( !empty( $field['fallback'] ) ) {
@@ -891,11 +971,11 @@ abstract class SiteOrigin_Widget extends WP_Widget {
 						$v_name = substr( $v_name, strpos($v_name, '][') + 2 );
 					}
 					$fallback_url = !empty( $all_values[$v_name . '_fallback'] ) ? $all_values[$v_name . '_fallback'] : '';
-
-					?><input type="text" value="<?php echo esc_url( $fallback_url ) ?>" placeholder="<?php esc_attr_e('Fallback Image URL', 'siteorigin-widgets') ?>" name="<?php echo $this->so_get_field_name( $name . '_fallback', $repeater ) ?>" class="media-fallback-external" /><?php
+					?><input type="text" value="<?php echo esc_url( $fallback_url ) ?>" placeholder="<?php esc_attr_e( 'External URL', 'siteorigin-widgets' ) ?>" name="<?php echo $this->so_get_field_name( $name . '_fallback', $repeater ) ?>" class="media-fallback-external siteorigin-widget-input" /><?php
 				}
-				?>
 
+				?>
+				<div class="clear"></div>
 				<input type="hidden" value="<?php echo esc_attr( is_array( $value ) ? '-1' : $value ) ?>" name="<?php echo $this->so_get_field_name( $name, $repeater ) ?>" class="siteorigin-widget-input" />
 				<?php
 				break;
@@ -1017,15 +1097,7 @@ abstract class SiteOrigin_Widget extends WP_Widget {
 				static $widget_font_families;
 				if( empty($widget_font_families) ) {
 
-					// Add the default fonts
-					$widget_font_families = array(
-						'Helvetica Neue' => 'Helvetica Neue',
-						'Lucida Grande' => 'Lucida Grande',
-						'Georgia' => 'Georgia',
-						'Courier New' => 'Courier New',
-					);
-
-					$widget_font_families = apply_filters('siteorigin_widgets_font_families', $widget_font_families );
+					$widget_font_families = siteorigin_widgets_font_families();
 				}
 				?>
 				<div class="siteorigin-widget-font-selector siteorigin-widget-field-subcontainer">
@@ -1060,7 +1132,14 @@ abstract class SiteOrigin_Widget extends WP_Widget {
 				break;
 
 			default:
-				?><?php _e('Unknown Field', 'siteorigin-widgets') ?><?php
+				// We couldn't find the field, so lets give other plugins a chance to provide it
+				echo apply_filters(
+					'siteorigin_widget_missing_field',
+					__('Unknown Field', 'siteorigin-widgets'),
+					$field,
+					$value,
+					$this
+				);
 				break;
 
 		}
@@ -1090,6 +1169,13 @@ abstract class SiteOrigin_Widget extends WP_Widget {
 		return $transformed;
 	}
 
+	/**
+	 * Convert a matched string to uppercase. Used as a preg callback.
+	 *
+	 * @param $matches
+	 *
+	 * @return string
+	 */
 	private function match_to_upper( $matches ) {
 		return strtoupper( $matches[1] );
 	}
@@ -1127,11 +1213,11 @@ abstract class SiteOrigin_Widget extends WP_Widget {
 	abstract function get_template_name($instance);
 
 	/**
-	 * Get the name of the directory in which we should look for the template.
+	 * Get the name of the directory in which we should look for the template. Relative to root of widget folder.
 	 *
 	 * @return mixed
 	 */
-	function get_template_dir($instance) {
+	function get_template_dir( $instance ) {
 		return 'tpl';
 	}
 
@@ -1172,7 +1258,6 @@ abstract class SiteOrigin_Widget extends WP_Widget {
 	 * @return object The instance
 	 */
 	function get_stored_instance($hash) {
-
 		return get_transient('sow_inst[' . $this->id_base . '][' . $hash . ']');
 	}
 
@@ -1210,12 +1295,16 @@ abstract class SiteOrigin_Widget extends WP_Widget {
 	}
 
 	/**
-	 * Can be overwritten by child widgets to make variables available to javascript via ajax calls.
+	 * Can be overwritten by child widgets to make variables available to javascript via ajax calls. These are designed to be used in the admin.
 	 */
-	function get_javascript_variables(){ }
+	function get_javascript_variables(){
+
+	}
 
 	/**
 	 * Used by child widgets to register scripts to be enqueued for the frontend.
+	 *
+	 * @param array $scripts an array of scripts. Each element is an array that corresponds to wp_enqueue_script arguments
 	 */
 	function register_frontend_scripts( $scripts ){
 		foreach ( $scripts as $script ) {
@@ -1225,6 +1314,9 @@ abstract class SiteOrigin_Widget extends WP_Widget {
 		}
 	}
 
+	/**
+	 * Enqueue all the registered scripts
+	 */
 	function enqueue_registered_scripts() {
 		foreach ( $this->frontend_scripts as $f_script ) {
 			if ( ! wp_script_is( $f_script[0] ) ) {
@@ -1241,6 +1333,8 @@ abstract class SiteOrigin_Widget extends WP_Widget {
 
 	/**
 	 * Used by child widgets to register styles to be enqueued for the frontend.
+	 *
+	 * @param array $styles an array of styles. Each element is an array that corresponds to wp_enqueue_style arguments
 	 */
 	function register_frontend_styles( $styles ) {
 		foreach ( $styles as $style ) {
@@ -1250,6 +1344,9 @@ abstract class SiteOrigin_Widget extends WP_Widget {
 		}
 	}
 
+	/**
+	 * Enqueue any frontend styles that were registered
+	 */
 	function enqueue_registered_styles() {
 		foreach ( $this->frontend_styles as $f_style ) {
 			if ( ! wp_style_is( $f_style[0] ) ) {
@@ -1279,8 +1376,4 @@ abstract class SiteOrigin_Widget extends WP_Widget {
 	 */
 	function enqueue_admin_scripts(){ }
 
-	/**
-	 * Initialize this widget in whatever way we need to. Run before rendering widget or form.
-	 */
-	function initialize(){ }
 }
